Made further improvements to the menu.
The menu behaves a lot more like other menus now, with full keyboard support, and mouse sweeping which allows you to switch to another active menu with the mouse without clicking. It was bothering me how janky the menu could be, so this work should make the menu component feel more pro.