Vastly improved the intellisense for JavaScript
Also hid the work I've been doing on TypeScript integration for now. It works ok, but Monaco has difficulty resolving paths for includes, and warns the user as such. I'd rather the editor fully work in that regard before making it available, so it'll remain disabled till I can figure out how to make that work.